+++ packages/core/src/components/Clickable/__stories__/Clickable.stories.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import Clickable, { ClickableProps } from "../Clickable";
 import Flex from "../../Flex/Flex";
 import { statesPlaySuite } from "../__tests__/Clickable.interactions";
-import "./Clickable.stories.scss";
+import Box from "../../Box/Box";
 
 export default {
   title: "Accessibility/Clickable",
@@ -11,8 +11,10 @@ export default {
 
 const clickableTemplate = (args: ClickableProps) => {
   return (
-    <Clickable className="monday-story-clickable_first-element" onClick={() => alert("clicked")} {...args}>
-      <div>I act like a button</div>
+    <Clickable onClick={() => alert("clicked")} {...args}>
+      <Box border padding="small" rounded="small">
+        I act like a button
+      </Box>
     </Clickable>
   );
 };
@@ -32,20 +34,15 @@ export const Overview = {
 export const States = {
   render: () => (
     <Flex gap="medium">
-      <Clickable
-        className="monday-story-clickable_first-element"
-        onClick={() => alert("clicked")}
-        ariaLabel="clickable button"
-      >
-        <div>Regular clickable element</div>
+      <Clickable onClick={() => alert("clicked")} ariaLabel="clickable button">
+        <Box border padding="small" rounded="small">
+          Regular clickable element
+        </Box>
       </Clickable>
-      <Clickable
-        className="monday-story-clickable_disabled-element"
-        onClick={() => alert("clicked")}
-        disabled
-        ariaLabel="disabled clickable button"
-      >
-        <div>Disabled clickable element</div>
+      <Clickable onClick={() => alert("clicked")} disabled ariaLabel="disabled clickable button">
+        <Box border backgroundColor="greyBackgroundColor" padding="small" rounded="small">
+          Disabled clickable element
+        </Box>
       </Clickable>
     </Flex>
   ),