+++ packages/core/src/components/SplitButton/__stories__/SplitButton.stories.tsx
@@ -1,7 +1,5 @@
 import React from "react";
-/* eslint-disable react/no-children-prop */
-import { createComponentTemplate } from "vibe-storybook-components";
-import SplitButton from "../SplitButton";
+import SplitButton, { SplitButtonProps } from "../SplitButton";
 import Button from "../../Button/Button";
 import { Add, Announcement, Check, Download, Favorite, Moon, Sun, Upload } from "@vibe/icons";
 import { createStoryMetaSettingsDecorator } from "../../../storybook";
@@ -9,7 +7,6 @@ import SplitButtonMenu from "../SplitButtonMenu/SplitButtonMenu";
 import MenuItem from "../../Menu/MenuItem/MenuItem";
 import Menu from "../../Menu/Menu/Menu";
 import MenuTitle from "../../Menu/MenuTitle/MenuTitle";
-import "./SplitButton.stories.scss";
 
 const metaSettings = createStoryMetaSettingsDecorator({
   component: SplitButton,
@@ -27,18 +24,20 @@ export default {
   },
 
   argTypes: metaSettings.argTypes,
-  decorators: metaSettings.decorators
+  decorators: metaSettings.decorators,
+  parameters: {
+    docs: {
+      liveEdit: {
+        scope: { Add, Announcement, Check, Download, Favorite, Moon, Sun, Upload }
+      }
+    }
+  }
 };
 
-const splitButtonTemplate = createComponentTemplate(SplitButton);
-
 export const Overview = {
-  render: splitButtonTemplate.bind({}),
-  name: "Overview",
+  render: (args: SplitButtonProps) => <SplitButton {...args}>Button</SplitButton>,
 
   args: {
-    children: "Button",
-
     secondaryDialogContent: () => (
       <SplitButtonMenu id="split-menu">
         <MenuItem icon={Check} title="Hey" />
@@ -57,16 +56,16 @@ export const Types = {
   render: () => (
     <>
       <SplitButton
-        children="Primary"
         secondaryDialogContent={
           <SplitButtonMenu id="split-menu">
             <MenuItem icon={Check} title="Hey" />
             <MenuItem icon={Announcement} title="There" />
           </SplitButtonMenu>
         }
-      />
+      >
+        Primary
+      </SplitButton>
       <SplitButton
-        children="Secondary"
         kind="secondary"
         secondaryDialogContent={
           <SplitButtonMenu id="split-menu">
@@ -74,9 +73,10 @@ export const Types = {
             <MenuItem icon={Announcement} title="There" />
           </SplitButtonMenu>
         }
-      />
+      >
+        Secondary
+      </SplitButton>
       <SplitButton
-        children="Tertiary"
         kind="tertiary"
         secondaryDialogContent={
           <SplitButtonMenu id="split-menu">
@@ -84,24 +84,17 @@ export const Types = {
             <MenuItem icon={Announcement} title="There" />
           </SplitButtonMenu>
         }
-      />
+      >
+        Tertiary
+      </SplitButton>
     </>
-  ),
-
-  name: "Types",
-
-  parameters: {
-    docs: {
-      liveEdit: { scope: { Check, Announcement } }
-    }
-  }
+  )
 };
 
 export const Sizes = {
   render: () => (
     <>
       <SplitButton
-        children="Small"
         size="small"
         secondaryDialogContent={
           <SplitButtonMenu id="split-menu">
@@ -109,9 +102,10 @@ export const Sizes = {
             <MenuItem icon={Announcement} title="There" />
           </SplitButtonMenu>
         }
-      />
+      >
+        Small
+      </SplitButton>
       <SplitButton
-        children="Medium"
         size="medium"
         secondaryDialogContent={
           <SplitButtonMenu id="split-menu">
@@ -119,9 +113,10 @@ export const Sizes = {
             <MenuItem icon={Announcement} title="There" />
           </SplitButtonMenu>
         }
-      />
+      >
+        Medium
+      </SplitButton>
       <SplitButton
-        children="Large"
         size="large"
         secondaryDialogContent={
           <SplitButtonMenu id="split-menu">
@@ -129,16 +124,11 @@ export const Sizes = {
             <MenuItem icon={Announcement} title="There" />
           </SplitButtonMenu>
         }
-      />
+      >
+        Large
+      </SplitButton>
     </>
-  ),
-
-  name: "Sizes",
-  parameters: {
-    docs: {
-      liveEdit: { scope: { Check, Announcement } }
-    }
-  }
+  )
 };
 
 export const SplitButtonWithIcons = {
@@ -168,19 +158,12 @@ export const SplitButtonWithIcons = {
       </SplitButton>
     </>
   ),
-
-  name: "Split button with icons",
-  parameters: {
-    docs: {
-      liveEdit: { scope: { Add, Check, Announcement } }
-    }
-  }
+  name: "Split button with icons"
 };
 
 export const SplitButtonAsThePrimaryAction = {
   render: () => (
     <SplitButton
-      children="Use template"
       secondaryDialogPosition="bottom-start"
       secondaryDialogContent={
         <SplitButtonMenu id="split-menu">
@@ -188,22 +171,17 @@ export const SplitButtonAsThePrimaryAction = {
           <MenuItem icon={Upload} title="Export template" />
         </SplitButtonMenu>
       }
-    />
+    >
+      Use template
+    </SplitButton>
   ),
-
-  name: "Split button as the primary action",
-  parameters: {
-    docs: {
-      liveEdit: { scope: { Download, Upload } }
-    }
-  }
+  name: "Split button as the primary action"
 };
 
 export const SecondarySplitButton = {
   render: () => (
     <>
       <SplitButton
-        children="Export to CSV"
         kind="secondary"
         secondaryDialogPosition="bottom-start"
         secondaryDialogContent={
@@ -213,11 +191,12 @@ export const SecondarySplitButton = {
             <MenuItem title="Export to Excel" />
           </SplitButtonMenu>
         }
-      />
+      >
+        Export to CSV
+      </SplitButton>
       <Button>New item</Button>
     </>
   ),
-
   name: "Secondary split button"
 };
 
@@ -226,7 +205,7 @@ export const CustomMenu = {
     <SplitButton
       secondaryDialogContent={
         <Menu focusItemIndexOnMount={2} id="menu" size="medium">
-          <MenuTitle caption="Look up, you might see" captionPosition={MenuTitle.positions.TOP} />
+          <MenuTitle caption="Look up, you might see" captionPosition="top" />
           <MenuItem icon={Sun} iconType="svg" title="The sun" />
           <MenuItem icon={Moon} iconType="svg" title="The moon" />
           <MenuItem icon={Favorite} iconType="svg" title="And the stars" />
@@ -236,11 +215,5 @@ export const CustomMenu = {
       Custom menu
     </SplitButton>
   ),
-
-  name: "Custom menu",
-  parameters: {
-    docs: {
-      liveEdit: { scope: { Sun, Moon, Favorite } }
-    }
-  }
+  name: "Custom menu"
 };