+++ packages/testkit/CHANGELOG.md
@@ -3,6 +3,61 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.12.0](https://github.com/mondaycom/vibe/compare/@vibe/testkit@1.11.2...@vibe/testkit@1.12.0) (2025-07-13)
+
+
+### Features
+
+* **testkit:** Add expand collapse component ([#2960](https://github.com/mondaycom/vibe/issues/2960)) ([9c701b1](https://github.com/mondaycom/vibe/commit/9c701b1e9536d551b3b060a03182db7ca3fae68e))
+
+
+
+
+
+## [1.11.2](https://github.com/mondaycom/vibe/compare/@vibe/testkit@1.11.1...@vibe/testkit@1.11.2) (2025-07-03)
+
+
+### Bug Fixes
+
+* update steps component to testkit ([#2936](https://github.com/mondaycom/vibe/issues/2936)) ([30923b1](https://github.com/mondaycom/vibe/commit/30923b1f2e3169856ea8bc2ef63b07ca077137fa))
+
+
+
+
+
+## [1.11.1](https://github.com/mondaycom/vibe/compare/@vibe/testkit@1.11.0...@vibe/testkit@1.11.1) (2025-07-01)
+
+
+### Bug Fixes
+
+* getText function in testkit ([#2959](https://github.com/mondaycom/vibe/issues/2959)) ([57b9437](https://github.com/mondaycom/vibe/commit/57b9437836594e0b485931164f94448e1c813e78))
+
+
+
+
+
+# [1.11.0](https://github.com/mondaycom/vibe/compare/@vibe/testkit@1.10.0...@vibe/testkit@1.11.0) (2025-06-19)
+
+
+### Features
+
+* **testkit:** common actions improvements ([#2930](https://github.com/mondaycom/vibe/issues/2930)) ([c0eb7be](https://github.com/mondaycom/vibe/commit/c0eb7be7aed6f78901e0011faefe0bb04e2d7cd9))
+
+
+
+
+
+# [1.10.0](https://github.com/mondaycom/vibe/compare/@vibe/testkit@1.9.2...@vibe/testkit@1.10.0) (2025-06-18)
+
+
+### Features
+
+* **Steps:** Add Steps component to testkit ([#2934](https://github.com/mondaycom/vibe/issues/2934)) ([0322f49](https://github.com/mondaycom/vibe/commit/0322f4944102921d54a64e922be4275bfcc42eba))
+
+
+
+
+
 ## [1.9.2](https://github.com/mondaycom/vibe/compare/@vibe/testkit@1.9.1...@vibe/testkit@1.9.2) (2025-05-21)